docs(README): add absurd coordinated omission results
`oha` definitely will have coordinated omission if I use the test
suggested in [[1]]. So, this adds the results if we still pretty much
uncap QPS and run the same thing again.

However, this makes Python look so bad I'm sure there is something wrong
still. For example, perhaps I should cap QPS lower for the python run
(e.g., 350) in order to not overwhelm FastAPI to such extremes. At the
same time, it seems more "fair" to run the same command for both.

[1]: https://redhatperf.github.io/post/coordinated-omission/